requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
Pillow>=10.0.0
